Colony and stack v3.51
Correction to move-insert courtesy of Peet Nick. Removal of end-block edge-case scenario code, as it was found to be of such low statistical significance as to be useless. Also increased memory use which may be it's own problem in some scenarios. To get around the edge case, if it occurs in a given context, an allocator approach would be most appropriate.