Fix another major bug with ring_iterator comparisons.
The expression `rv.begin() == rv.end()` should evaluate true only if
`rv.size() == 0`; similarly, `rv.begin() <= rv.end()` should always
evaluate true. The use of `modulo_capacity` in the comparison operators
was just plain wrong; and as that was the *only* use of `modulo_capacity`,
we can and should get rid of that method.