Remove an "explicit" on copy_popper's constructor that fails a test.
I think I added both the "explicit" and the test relying on the
non-explicit constructor, in two different branches. :P
I tend to think that "explicit" is good and so the fix should be
the other way around, but this fix preserves the status quo and fixes
the tests so I'm committing the fix directly.