slot_map: Adjust "reserve" and "capacity" member function API.
The member function `sm.reserve_slots(n)` used to call `reserve` (and nothing
else) on the slots container. This is not useful functionality. After this
patch, `sm.reserve_slots(n)` resizes the slots container to `n`. It does not
touch the values container, since there is a valid use-case for "many slots,
relatively few values."

The member function `sm.reserve(n)` used to call `reserve` on all three
containers. After this patch, `sm.reserve(n)` still calls `reserve` on the
values container, and also on the reverse_map container, which is always
the same size as the values container by definition. It now *resizes*
the slots container to `n`, since there is no way for the programmer to
use the `n` capacity of the values container without taking up `n` slots
anyway.

N.B.: For container types that do not provide a suitable `reserve()` method,
the methods `sm.reserve(n)` and `sm.reserve_slots(n)` have exactly the same
effects.

The accessor `sm.capacity_slots()` has been renamed to `sm.slot_count()`
for clarity, and returns the `size()` of the slots container.

The accessor `sm.capacity()` still returns the `capacity()` of the
values container. It now SFINAEs away if the container does not provide
a suitable `capacity()` method.