inplace_function: Test various qualifications of `nullptr`.
I was worried about some of these construct-from-qualified-nullptr
cases, but they all work fine! (And in fact I was not able to break
any of them by introducing bugs into `inplace_function`, which
indicates that technically these new tests are redundant. But it's
nice to document that this case has been attempted, and works.)