Remove the overly complicated interactive test from "ring_test.cpp".
This is an alternative solution to what @p-groake proposed in #137.
Simply removing the test is cleaner: the test doesn't hit any codepath
that isn't hit by other tests already, and it drags in a lot of
non-SG14 stuff (`future`, `iostream`), and even if we went with
the proposal in #137, its behavior would be non-deterministic, which
is not good for CI. So, removed.