Stop using std::aligned_storage. Stop using std::alignment_of.
The former is wrong more often than it's right, so let's just
switch to our hand-rolled, 100% correct implementation and get rid
of the hard-to-maintain ifdefs.

The latter is obsolete compared to the `alignof` keyword, and
just ends up making lots of template instantiations that we
don't need to be paying for.