inplace_function: Remove noexcept-specifications from lambdas where it doesn't matter.
These lambdas are immediately stored into function pointers which are *not*
declared `noexcept`, so the fact that the lambda itself is `noexcept` is not
going to help anyone. (Except compilers sufficiently smart to "devirtualize"
the call through the pointer and realize that the pointer always points to the
lambda. But in that hypothetical case, such a compiler is probably smart enough
to see that the lambda's body is noexcept, without needing "help" from us.)

This is a followup to my previous commit, which fixed a bug in one noexcept
specification. Even better than fixing buggy code is *removing* buggy code!