inplace_function: Use the "assign by value" idiom. NFC.
As the untouched unit tests demonstrate, this doesn't (AFAIK)
change any of the observable behavior of `inplace_function`;
but it drastically simplifies the code, because `operator=`
no longer calls any user-provided functions that are allowed
to throw.

(Storing a wrapped object which throws on move-construction
or throws on destruction causes undefined behavior, as far as
we're concerned. The usual disclaimer applies: we promise not
to force you to write `noexcept` in your code, as long as you
promise you won't throw in practice regardless.)