Squashed commit of the `flat_set` and `flat_map` branches.
This is a complete implementation of P1222 `flat_set` and
P0429 `flat_map`, but with my own bugfixes and interpretations
of certain places that the wording paper deviates from consistency
or common sense.

For example, `swap` is `noexcept`. The wording paper tries to
make it SFINAE-constrained and/or conditionally-noexcept. But
really, swap just IS noexcept; we shouldn't be innovating in this
area.

For another example, constructing a `flat_set` from a `KeyContainer`
should not cause _two_ copies of the container's elements. (This has
not been completely fixed, e.g. in `flat_map`.)

I encourage people to try it out, open issues, and make pull requests.

The most recent public wording papers are here:
http://www.open-std.org/jtc1/sc22/wg21/docs/papers/2019/p1222r1.pdf
http://www.open-std.org/jtc1/sc22/wg21/docs/papers/2019/p0429r6.pdf

Zach has minorly updated versions on the Cologne wiki:
http://wiki.edg.com/pub/Wg21cologne2019/LibraryWorkingGroup/D1222R2.pdf
http://wiki.edg.com/pub/Wg21cologne2019/LibraryWorkingGroup/D0429R7.pdf