flat_set: Eliminate the last C++14isms.
No more generic lambdas in std::partition_point helper functions.
In fact, it occurred to me that these generic lambdas were perhaps
technically incorrect, because they result in our passing proxy
reference types into the comparator when really we should be passing
const references to the `Key` type. So, add an evil test for that.